new folder structure for tx-builder form
