+++ apps/tx-builder/src/components/forms/fields/TextareaContractField.tsx
@@ -1,10 +1,10 @@
 import { TextFieldInputProps } from '@gnosis.pm/safe-react-components/dist/inputs/TextFieldInput';
-import ContractTextField from './ContractTextField';
+import TextContractField from './TextContractField';
 
 const DEFAULT_ROWS = 4;
 
 function TextareaContractField(props: TextFieldInputProps) {
-  return <ContractTextField {...props} multiline rows={DEFAULT_ROWS} />;
+  return <TextContractField {...props} multiline rows={DEFAULT_ROWS} />;
 }
 
 export default TextareaContractField;
