+++ apps/tx-builder/src/components/forms/validations/validateAddressField.ts
@@ -1,0 +1,9 @@
+import { isValidAddress } from '../../../utils';
+
+function validateAddressField(value: string): string | undefined {
+  if (!isValidAddress(value)) {
+    return 'Invalid Address';
+  }
+}
+
+export default validateAddressField;
