+++ apps/tx-builder/src/components/forms/validations/validateAmountField.ts
@@ -1,0 +1,9 @@
+import { isInputValueValid } from '../../../utils';
+
+function validateAmountField(value: string): string | undefined {
+  if (!isInputValueValid(value)) {
+    return 'Invalid Amount';
+  }
+}
+
+export default validateAmountField;
